Removed all cruff in inputs, preserved use of namedtuple
