+++ tests/conftest.py
@@ -986,102 +986,4 @@ def setup_minimal_water(setup_folder_structure,
     return str(setup_folder_structure)
 
 
-@pytest.fixture(scope='function')
-def one_input():
-    """Returns a model input dictionary for the example water model with one
-    decision variable and one parameter
-    """
-    inputs = {
-        'decision variables': [
-            {
-                'name': 'water treatment capacity',
-                'bounds': (0, 20),
-                'value': 10
-            }
-        ],
-        'parameters': [
-            {
-                'name': 'raininess',
-                'bounds': (0, 5),
-                'value': 3
-            }
-        ]
-    }
 
-    return inputs
-
-
-@pytest.fixture(scope='function')
-def two_inputs():
-    """Returns a model input dictionary for the example water model with two
-    decision variables and one parameter
-    """
-    inputs = {
-        'decision variables': [
-            {
-                'name': 'reservoir pumpiness',
-                'bounds': (0, 100),
-                'value': 24.583
-            },
-            {
-                'name': 'water treatment capacity',
-                'bounds': (0, 20),
-                'value': 10
-            }
-        ],
-        'parameters': [
-            {
-                'name': 'raininess',
-                'bounds': (0, 5),
-                'value': 3
-            }
-        ]
-    }
-    return inputs
-
-
-@pytest.fixture(scope='function')
-def one_dependency():
-    """Returns a model input dictionary with a single (unlikely to be met)
-    dependency
-    """
-    inputs = {
-        'dependencies': [
-            {
-                'name': 'macguffins produced',
-                'spatial_resolution': 'LSOA',
-                'temporal_resolution': 'annual',
-                'from_model': 'macguffins_model'
-            }
-        ]
-    }
-    return inputs
-
-
-@pytest.fixture(scope='function')
-def dynamic_data():
-    """Returns a model input dictionary for the example water model with two
-    decision variables and one parameter
-    """
-    inputs = {
-        'decision variables': [
-            {
-                'name': 'new capacity',
-                'bounds': (0, 5),
-                'value': 10
-            },
-        ],
-        'parameters': [
-            {
-                'name': 'raininess',
-                'bounds': (0, 5),
-                'value': 3
-            },
-            {
-                'name': 'existing capacity',
-                'bounds': (0, 999),
-                'value': 1
-            }
-        ]
-    }
-    return inputs
